+++ newdle/client/src/components/Home.js
@@ -1,11 +1,22 @@
 import React from 'react';
-import LoginLink from './LoginLink';
+import {Button} from 'semantic-ui-react';
+import LoginRequired from './LoginRequired';
+import {useRouter} from '../util/router';
 
 export default function Home() {
+  const router = useRouter();
+
   return (
     <>
       <h2>Home</h2>
-      <LoginLink to="/new">Create newdle</LoginLink>
+      <LoginRequired
+        component={Button}
+        onClick={() => {
+          router.history.push('/new');
+        }}
+      >
+        Create newdle
+      </LoginRequired>
     </>
   );
 }
